Fix some errors and warnings

+++ src/enums.h
@@ -305,7 +305,7 @@ struct enum_traits<reachability_cache_quadrant> {
     static constexpr int size = static_cast<int>( last ) + 1;
 
     inline static reachability_cache_quadrant quadrant( bool S, bool W ) {
-        return static_cast<reachability_cache_quadrant>( ( W << 1 ) | S );
+        return static_cast<reachability_cache_quadrant>( ( W << 1 ) | static_cast<int>( S ) );
     }
 };
